Suppression de la fonction random_possibilities au profit de la fonction getRandomMoves() donnée dans la version corriger de move.c Correction de la fonction getRandomMoves() pour qu'il ne prend pas toujours le même seed de random Séparation des fonctions dans plusieurs librairies pour plus de lisibilité
