+++ moves.c
@@ -165,6 +165,7 @@ void updateLocalisation(t_localisation *p_loc, t_move m)
 
 t_move *getRandomMoves(int N)
 {
+    srand(time(NULL));  //Set la fonction rand
     int nbmoves[]={22,15,7,7,21,21,7};
     int total_moves=100;
     t_move *moves = (t_move *)malloc(N * sizeof(t_move));
